[Vt] Fix signed/unsigned comparison warning.

+++ pxr/base/lib/vt/operators.h
@@ -167,7 +167,7 @@
             return VtArray<T>();                                             \
         }                                                                    \
         VtArray<rettype> ret(vec.size());                                \
-        for (int i = 0; i < length; ++i) {                                   \
+        for (size_t i = 0; i < length; ++i) {                                   \
             if (!extract<T>(obj[i]).check())                              \
                 TfPyThrowValueError("Element is of incorrect type.");        \
             ret[i] = expr;                                                   \
@@ -226,7 +226,7 @@
             return VtArray<bool>();                                   \
         }                                                             \
         VtArray<bool> ret(vec.size());                            \
-        for (int i = 0; i < length; ++i) {                            \
+        for (size_t i = 0; i < length; ++i) {                            \
             if (!extract<T>(obj[i]).check())                       \
                 TfPyThrowValueError("Element is of incorrect type."); \
             ret[i] = expr;                                            \
